# api/swagger.yaml
openapi: 3.0.1
info:
  title: Registration
  description: Registration global service
  termsOfService: http://swagger.io/terms/
  contact:
    email: agnetiuslee@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: https://usahakreatifindonesia.co.id/register
security:
- secret_key: []
tags:
- name: producer
  description: Producer Message 
- name: consumer
  description: Consumer Message
basePath: "/thirdparty"
paths:
  
  /google/email:
    get:
      tags:
      - "get email"
      summary: "Get email"
      operationId: "getGoogleEmail"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "employeeId"
        in: "header"
        required: true
        type: "string"
      - name: "param"
        in: "header"
        required: true
        type: "string"
        description: '{"id": "", "maxResults": "10", "q": "subject:refund", "labelIds": "UNREAD"}'
      # - in: "body"
      #   name: "body"
      #   required: true
      #   schema:
      #     $ref: "#/definitions/getPresentationSchedule"
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "google"
    
    put:
      tags:
      - "update email"
      summary: "update email"
      operationId: "updateGoogleEmail"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "employeeId"
        in: "header"
        required: true
        type: "string"
      # - name: "param"
      #   in: "header"
      #   required: false
      #   type: "string"
      #   description: '{"id": "", "maxResults": "10", "q": "subject:refund", "labelIds": "UNREAD"}'
      requestBody:
        content:
          '*/*':
            schema:
              required: true
              $ref: '#/components/schemas/updateGoogleEmail'
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "google"

  /google/calendar:
    post:
      tags:
      - "Google add calendar event"
      summary: "Create New Schedule of Presentation"
      operationId: "createGoogleCalendarEvent"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "header"
        required: true
        type: "string"        
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/createGoogleCalendarEvent'
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "google"
    
    get:
      tags:
      - "Presentation Schedule"
      summary: "Get All Schedule of Presentation"
      operationId: "getGoogleCalendarEvent"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "header"
        required: true
        type: "string"
      - name: "param"
        in: "header"
        required: true
        type: "string"
        description: '{ "maxResult": "10" }'
      # - in: "body"
      #   name: "body"
      #   required: true
      #   schema:
      #     $ref: "#/definitions/getPresentationSchedule"
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "google"

  /google/maps:
    get:
      tags:
      - "google get lat long"
      summary: "google get long"
      operationId: "getAddressCoordinate"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "header"
        required: true
        type: "string"
      - name: "param"
        in: "header"
        required: true
        type: "string"
        description: '{ "address": "jl. tanjung wira no 19 gresik" }'
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "google"

  /google/auth:
    post:
      tags:
      - "Google code for save token"
      summary: "Create New Schedule of Presentation"
      operationId: "createGoogleToken"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "employeeId"
        in: "header"
        required: true
        type: "string"
      requestBody:
        content:
          '*/*':
            schema:
              required: true
              $ref: '#/components/schemas/createGoogleToken'
        required: true        
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "google"

    get:
      tags:
      - "google check auth"
      summary: "google check auth"
      operationId: "checkGoogleToken"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "employeeId"
        in: "header"
        required: true
        type: "string"
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "google"

  /trello/auth:
    get:
      tags:
      - "google check auth"
      summary: "google check auth"
      operationId: "getAuthData"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "employeeId"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "trello"
    post:
      tags:
      - "Google code for save token"
      summary: "Create New Schedule of Presentation"
      operationId: "postAuthData"
      produces:
      - "application/xml"
      - "application/json"
      parameters:

      requestBody:
        content:
          '*/*':
            schema:
              required: true
              $ref: '#/components/schemas/createToken'
        required: true        
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "trello"

  /trello/board:
    get:
      tags:
      - "google check auth"
      summary: "google check auth"
      operationId: "getBoard"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "employeeId"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "trello"
    put:
      tags:
      - "google check auth"
      summary: "google check auth"
      operationId: "updateBoard"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      requestBody:
        content:
          '*/*':
            schema:
              required: false
              $ref: '#/components/schemas/trelloBoard'
        required: true
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "trello"
    delete:
      tags:
      - "google check auth"
      summary: "google check auth"
      operationId: "deleteBoard"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      requestBody:
        content:
          '*/*':
            schema:
              required: false
              $ref: '#/components/schemas/trelloBoard'
        required: true
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "trello"

  /trello/list:
    get:
      tags:
      - "google check auth"
      summary: "google check auth"
      operationId: "getList"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "employeeId"
        in: "query"
        required: true
        type: "string"
      - name: "boardId"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "trello"
    put:
      tags:
      - "google check auth"
      summary: "google check auth"
      operationId: "updateList"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      requestBody:
        content:
          '*/*':
            schema:
              required: false
              $ref: '#/components/schemas/trelloBoard'
        required: true
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "trello"
    delete:
      tags:
      - "google check auth"
      summary: "google check auth"
      operationId: "deleteList"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      requestBody:
        content:
          '*/*':
            schema:
              required: false
              $ref: '#/components/schemas/trelloBoard'
        required: true
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "trello"

  /trello/card:
    get:
      tags:
      - "google check auth"
      summary: "google check auth"
      operationId: "getCard"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "employeeId"
        in: "query"
        required: true
        type: "string"
      - name: "listId"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "trello"
    put:
      tags:
      - "google check auth"
      summary: "google check auth"
      operationId: "updateCard"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      requestBody:
        content:
          '*/*':
            schema:
              required: false
              $ref: '#/components/schemas/trelloBoard'
        required: true
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "trello"
    delete:
      tags:
      - "google check auth"
      summary: "google check auth"
      operationId: "deleteCard"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      requestBody:
        content:
          '*/*':
            schema:
              required: false
              $ref: '#/components/schemas/trelloBoard'
        required: true
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "trello"

  /trello/sync:
    get:
      tags:
      - "google check auth"
      summary: "google check auth"
      operationId: "getSync"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "params"
        in: "header"
        required: true
        type: "string"
      responses:
        200:
          description: "request success"
        406:
          description: "request"
        500:
          description: "internal server error"
      x-swagger-router-controller: "trello"
components:
  schemas:
    updateGoogleEmail:
      type: "object"
      properties:
        # employeeId:
        #   type: "string"
        #   example: "1,2,3"
        id:
          type: "string"
          example: "1747707c807ed8ac"
        removeLabelIds:
          type: "array"
          items:
            type: string
          example: "['UNREAD']"

    createGoogleCalendarEvent:
      type: "object"
      properties:
        # employeeId:
        #   type: "string"
        #   example: "1,2,3"
        summary:
          type: "string"
          example: "Meeting with: Muhamad Syafriansyah"
        place:
          type: "string"
          example: "Pakuwon Center #16-05, Jl. Embong Malang No.4, Kedungdoro, Tegalsari, Surabaya City, East Java 60261"
        description:
          type: "string"
          example: "A chance to hear more about products."
        date:
          type: "string"
          example: "2020-04-15"
        time:
          type: "string"
          example: "10:04"
        duration:
          type: "string"
          example: "1, 2, 3 (in hour)"
        timeZone:
          type: "string"
          example: "Asia/Jakarta"
        email:
          type: "string"
          example: "safriansah@gmail.com"
        
    createGoogleToken:
      type: "object"
      properties:
        # employeeId:
        #   type: "string"
        #   example: "1,2,3"
        code:
          type: "string"
          example: "4/ygFmQxktObX5LqwBm6n-_FAqysR_vbK5ZcUgiFoSf5NTgikMM0XqgQfBeaU64Lp_m5Srz0v21s_ZXw03kTYDnuM"

    createToken:
      type: "object"
      properties:
        employeeId:
          type: "string"
          example: "1,2,3"
        token:
          type: "string"
          example: "4/ygFmQxktObX5LqwBm6n-_FAqysR_vbK5ZcUgiFoSf5NTgikMM0XqgQfBeaU64Lp_m5Srz0v21s_ZXw03kTYDnuM"
    
    trelloBoard:
      type: "object"
      properties:
        employeeId:
          type: "string"
        boardId:
          type: "string"
        name:
          type: "string"

    # consumerSchema:
    #   type: object
    #   properties:
    #     content:
    #       type: string
  # securitySchemes:
  #   petstore_auth:
  #     type: oauth2
  #     flows:
  #       implicit:
  #         authorizationUrl: http://petstore.swagger.io/oauth/dialog
  #         scopes:
  #           write:pets: modify pets in your account
  #           read:pets: read your pets
  #   secret_key:
  #     type: apiKey
  #     name: SecretKey
  #     in: header
  #   oauth_auth:
  #     type: oauth2
  #     flows:
  #       implicit:
  #         authorizationUrl: https://www.linkedin.com/in/yohanes-ronald-agnetius-sengkey-lee-b174a940/
  #         scopes:
  #           read:auth: read authentication
  #   state:
  #     type: apiKey
  #     name: state
  #     in: header